[Refactor] Create SupportContextNavigator and inject with bind to SmartChecklistNavigation
